% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContSimmap_fns.R
\name{make_simmaps_BePhyNE}
\alias{make_simmaps_BePhyNE}
\title{Simulate continuous maps from BePhyNE posterior samples}
\usage{
make_simmaps_BePhyNE(tree, logdf, char_names = NA, nsims = 10)
}
\arguments{
\item{tree}{A phylo object.}

\item{logdf}{BePhyNE MCMC log data frame returned by [read_BePhyNE_log()].}

\item{char_names}{Optional character names for the simulated continuous
characters. If `NA`, names are generated from predictor indices.}

\item{nsims}{Number of posterior samples/maps to simulate.}
}
\value{
A list of continuous stochastic maps produced by
  `contsimmap::make.contsimmap()`.
}
\description{
Converts sampled BePhyNE posterior parameters into continuous stochastic maps
for niche optimum and breadth traits.
}
