% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_fns.R
\name{plot_BePhyNE_simmap}
\alias{plot_BePhyNE_simmap}
\title{Plot BePhyNE continuous stochastic maps}
\usage{
plot_BePhyNE_simmap(contsimmaps, scale_atr = NA)
}
\arguments{
\item{contsimmaps}{Continuous stochastic maps returned by
[make_simmaps_BePhyNE()].}

\item{scale_atr}{Optional scaling attributes from [format_BePhyNE_data()] used
to label or back-transform mapped values where implemented.}
}
\value{
Called for its plotting side effect.
}
\description{
Plots continuous stochastic maps generated by [make_simmaps_BePhyNE()].
}
