% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process_MCMC.R
\name{read_BePhyNE_log}
\alias{read_BePhyNE_log}
\title{Read a BePhyNE parameter log}
\usage{
read_BePhyNE_log(file_name)
}
\arguments{
\item{file_name}{Path to the BePhyNE log file returned by [BePhyNE_MCMC()].}
}
\value{
A data frame containing sampled MCMC parameter values.
}
\description{
Loads a tab-delimited `.pars.log` file produced by [BePhyNE_MCMC()].
}
