% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model_adequacy_fns.R
\name{separate.data}
\alias{separate.data}
\title{Split BePhyNE data into training and prediction sets}
\usage{
\method{separate}{data}(data_final, ratio = 0.5)
}
\arguments{
\item{data_final}{A BePhyNE species data list.}

\item{ratio}{Proportion assigned to training.}
}
\value{
A list with `training` and `predicting`.
}
\description{
Split BePhyNE data into training and prediction sets
}
